+++ hbw/config/initializers/bpm.rb
@@ -1,0 +1,5 @@
+require_relative '../../lib/bpm/config_validator'
+
+Rails.application.config.to_prepare do
+  HBW::Common::API.config = BPM::ConfigValidator.(config: Settings::BPM[Rails.env])
+end