+++ hbw/app/controllers/hbw/events/tasks_controller.rb
@@ -3,11 +3,12 @@ module HBW
     class TasksController < APIController
       def update
         event = {
-          task_id:  params['id'],
-          name:     params['event_name'],
-          assignee: params['assignee'],
-          version:  params['version'],
-          users:    params['users'] || []
+          task_id:     params['id'],
+          name:        params['event_name'],
+          assignee:    params['assignee'],
+          version:     params['version'],
+          users:       params['users'] || [],
+          process_key: params['process_key']
         }
 
         HBW::TaskNotifier.(widget:, event:)