+++ hbw/app/javascript/packs/hbw/components/form/select.js.jsx
@@ -197,7 +197,9 @@ modulejs.define(
       const renderDescription = () => {
         const { placement, text } = params.description;
 
-        return <div className="description" data-test={`description-${placement}`}>{text}</div>;
+        const translated = translateBP(`${task.process_key}.${task.key}.${name}.description`, {}, text);
+
+        return <div className="description" data-test={`description-${placement}`}>{translated}</div>;
       };
 
       const buildOptions = () => {