+++ CHANGELOG.md
@@ -49,6 +49,11 @@ v2.9 [Unreleased]
 ### Breaking changes
 - [#719](https://github.com/hydra-billing/homs/pull/719) Use REDIS_HOST and REDIS_PORT env variables instead of REDIS_URL for actioncable.
 
+v2.8.12 [Unreleased]
+-------------------
+### Features
+- [#757](https://github.com/hydra-billing/homs/pull/757) Allow to localize description for BP form fields.
+
 v2.8.11 [2024-08-12]
 -------------------
 ### Features